Update client.lua
Sleeper on line 61 increased to 7 over 5 , The sleeper at 5 was causing 10MS when in range of the object we were trying to get distance from, The sleeper at 7 keeps this in a more manageable range of 5-6ms. https://prnt.sc/rj7hot This is just a cheap fix. Also using Vdist2 over Vdist as we dont need any sqrt in these checks for objects, so Vdist2 is actually a faster native.
Launder will need a fix too as it spikes to 15ms.
Rig is an I9-9900k and 2080ti , so performance spikes this high will be detrimental to lower end hardware if people just start mashing scripts together.