Update server.lua
This makes a functional Start Work event
When passed from another server file IE > police has a marker , you walk into the marker to start work , the file that police marker is in holds an event that defines the job as well as src information as for some reason its not being retrieved right in jobcore.
So I do 
>>> RegisterServerEvent('test1')
AddEventHandler('test1', function()
    local jobTitle = "SHERIFF"
    local src = source
    TriggerEvent('DRP_Jobs', src, jobTitle)
end) 
Ill let you work that part out, this is just a bandaid fix and should probably be treated as such. But this will give you information as to why its not working for the rest of community.